Initial prototype for remote query cache plugin. Re-factored the common code for CachedResultSet.
